+++ cli/src/tool/tool_validate.rs
@@ -86,6 +86,8 @@ async fn validate_off_chain_tool(url: reqwest::Url) -> AnyResult<ToolMeta, Nexus
         )));
     }
 
+    // TODO: <https://github.com/Talus-Network/nexus-sdk/issues/107>
+
     meta_handle.success();
 
     Ok(meta)